Add configuration tuning
Currently this only tunes blood groups. It does this through the
TuneConfig function, called as TuneConfig(config, iters, maxError).
This function will create an instance, then check the distributions of
blood types in the resulting donors and recipients independently. It
then calculates the largest discrepancy, and adjusts the probability
distributions of each blood type among donors and recipients, and
repeats the process.

This ends once the error is small enough, or a maximum number of
iterations has completed. Note that a large number of iterations is not
guaranteed to converge, especially if the number of recipients is small.
I'd recomment tuning with upwards of 2000 recipients, then creating
instances based on any resulting probabilities.