Add support for multiple PRA distributions
There is a high correlation (with p <<< 0.01) between a recipient having
a blood compatible donor, and the recipients cPRA level. This adds
support for "compatPraBandsString" and "incompatBraBandsString", which
can hold cPRA bands for the distributions of cPRA levels for recipients
with and without a blood compatible donor, respectively.

Note that this will fall back to the regular praBandsString if those are
not present.